# src/main/resources/REST-API-DOC-swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: Phone Shop REST API Interface
    description: Phone Shop REST API Interface
  - name: basic-error-controller
    description: Basic Error Controller
paths:
  /error:
    get:
      tags:
        - basic-error-controller
      summary: error
      operationId: errorUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties:
              type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    head:
      tags:
        - basic-error-controller
      summary: error
      operationId: errorUsingHEAD
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties:
              type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags:
        - basic-error-controller
      summary: error
      operationId: errorUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties:
              type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - basic-error-controller
      summary: error
      operationId: errorUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties:
              type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - basic-error-controller
      summary: error
      operationId: errorUsingDELETE
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties:
              type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    options:
      tags:
        - basic-error-controller
      summary: error
      operationId: errorUsingOPTIONS
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties:
              type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - basic-error-controller
      summary: error
      operationId: errorUsingPATCH
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties:
              type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /phoneshop/products:
    get:
      tags:
        - Phone Shop REST API Interface
      summary: Get a list of all products stored in application.
      description: Retrieve all products details.
      operationId: fetchAllProductUsingGET
      produces:
        - '*/*'
        - text/plain
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProductCatalog'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - Phone Shop REST API Interface
      summary: Receive a product details as a request to be created on Phone Shop.
      description: Receive an request for create a new product in store and return the product with Id.
      operationId: createProductUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
        - text/plain
      parameters:
        - in: body
          name: request
          description: request
          required: true
          schema:
            $ref: '#/definitions/Item'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/phoneshop/products/{productId}':
    put:
      tags:
        - Phone Shop REST API Interface
      summary: Receive a product details as an updateRequest to be updated on Phone Shop.
      description: Receive an updateRequest for update an existing product in store.
      operationId: updateProductUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
        - text/plain
      parameters:
        - name: productId
          in: path
          description: productId
          required: true
          type: integer
          format: int32
        - in: body
          name: updateRequest
          description: updateRequest
          required: true
          schema:
            $ref: '#/definitions/Item'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /phoneshop/purchase:
    post:
      tags:
        - Phone Shop REST API Interface
      summary: Receive a list of products as an orderRequest to buy from the  Phone Shop.
      description: Recieve an orderRequest for buying a list of products in store and will receive an invoice.
      operationId: purchaseUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
        - text/plain
      parameters:
        - in: body
          name: orderRequest
          description: orderRequest
          required: true
          schema:
            $ref: '#/definitions/Create Purchase OrderRequest'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /phoneshop/types:
    get:
      tags:
        - Phone Shop REST API Interface
      summary: Get a list of all product types stored in application.
      description: Retrieve all product types details.
      operationId: fetchAllProductTypesUsingGET
      produces:
        - '*/*'
        - text/plain
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProductCatalog'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - Phone Shop REST API Interface
      summary: Receive a product type details as a request to create a ProductType on Phone Shop.
      description: Receive an request for create a new product type in store and return the ProductType with Id.
      operationId: createProductTypeUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
        - text/plain
      parameters:
        - in: body
          name: request
          description: request
          required: true
          schema:
            $ref: '#/definitions/Item'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/phoneshop/types/{typeId}':
    put:
      tags:
        - Phone Shop REST API Interface
      summary: Receive a product type details to update a ProductType on Phone Shop.
      description: Receive an updateRequest for update an existing ProductType in store.
      operationId: updateProductUsingPUT_1
      consumes:
        - application/json
      produces:
        - '*/*'
        - text/plain
      parameters:
        - name: typeId
          in: path
          description: typeId
          required: true
          type: integer
          format: int32
        - in: body
          name: updateRequest
          description: updateRequest
          required: true
          schema:
            $ref: '#/definitions/Item'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  Create Purchase OrderRequest:
    type: object
    required:
      - items
    properties:
      items:
        type: array
        description: A collection of selected product Id's
        allowEmptyValue: false
        items:
          type: integer
          format: int32
    title: Create Purchase OrderRequest
    description: OrderRequest Request resource representation
  Item:
    type: object
    properties:
      category:
        type: string
      desc:
        type: string
      id:
        type: integer
        format: int32
      price:
        type: number
        format: double
      productName:
        type: string
      quantity:
        type: integer
        format: int32
      typeName:
        type: string
    title: Item
    description: Item resource representation
  ModelAndView:
    type: object
    properties:
      empty:
        type: boolean
      model:
        type: object
      modelMap:
        type: object
        additionalProperties:
          type: object
      reference:
        type: boolean
      status:
        type: string
        enum:
          - '100'
          - '101'
          - '102'
          - '103'
          - '200'
          - '201'
          - '202'
          - '203'
          - '204'
          - '205'
          - '206'
          - '207'
          - '208'
          - '226'
          - '300'
          - '301'
          - '302'
          - '303'
          - '304'
          - '305'
          - '307'
          - '308'
          - '400'
          - '401'
          - '402'
          - '403'
          - '404'
          - '405'
          - '406'
          - '407'
          - '408'
          - '409'
          - '410'
          - '411'
          - '412'
          - '413'
          - '414'
          - '415'
          - '416'
          - '417'
          - '418'
          - '419'
          - '420'
          - '421'
          - '422'
          - '423'
          - '424'
          - '426'
          - '428'
          - '429'
          - '431'
          - '451'
          - '500'
          - '501'
          - '502'
          - '503'
          - '504'
          - '505'
          - '506'
          - '507'
          - '508'
          - '509'
          - '510'
          - '511'
      view:
        $ref: '#/definitions/View'
      viewName:
        type: string
    title: ModelAndView
  ProductCatalog:
    type: object
    required:
      - productCatalog
    properties:
      productCatalog:
        type: array
        description: A collection of the returned product details
        allowEmptyValue: false
        items:
          $ref: '#/definitions/Item'
    title: ProductCatalog
    description: Product Catalog resource representation
  ResponseEntity:
    type: object
    properties:
      body:
        type: object
      statusCode:
        type: string
        enum:
          - '100'
          - '101'
          - '102'
          - '103'
          - '200'
          - '201'
          - '202'
          - '203'
          - '204'
          - '205'
          - '206'
          - '207'
          - '208'
          - '226'
          - '300'
          - '301'
          - '302'
          - '303'
          - '304'
          - '305'
          - '307'
          - '308'
          - '400'
          - '401'
          - '402'
          - '403'
          - '404'
          - '405'
          - '406'
          - '407'
          - '408'
          - '409'
          - '410'
          - '411'
          - '412'
          - '413'
          - '414'
          - '415'
          - '416'
          - '417'
          - '418'
          - '419'
          - '420'
          - '421'
          - '422'
          - '423'
          - '424'
          - '426'
          - '428'
          - '429'
          - '431'
          - '451'
          - '500'
          - '501'
          - '502'
          - '503'
          - '504'
          - '505'
          - '506'
          - '507'
          - '508'
          - '509'
          - '510'
          - '511'
      statusCodeValue:
        type: integer
        format: int32
    title: ResponseEntity
  View:
    type: object
    properties:
      contentType:
        type: string
    title: View